miniapp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
python-multipart==0.0.12